Fixed a CUDA out of memory bug
Fixed a CUDA out of memory bug when detecting with a custom model.  The model was loaded onto the GPU when getting the custom labels, however it was loaded again during actual detection, resulting on double the memory use. The model is no only loaded on the CPU when getting the custom labels, leaving the GPU free form the detection. model. No slow down as this is not a GPU required operation.